+++ demo/demo_cgal.py
@@ -92,7 +92,7 @@ basis_options = BasisOptimizerOptions(
     do_snapping=True,
     do_forcing=False,
     max_iterations=10,
-    lambda_smooth=0.1,
+    lambda_smoothing=0.1,
 )
 
 fit_options = FitOptions(